chore(structure): rename folders and add demo video to docs
