--- src/adapte/book-adapter-interface.ts
@@ -1,6 +1,0 @@
-export interface IBookAdapter<Entity, Dto> 
-{
-  entityToDto(model: Entity): Dto;
-
-  dtoToEntity(dto: Dto): Promise<Entity>;
-}
