+++ src/adapter/book-adapter-interface.ts
@@ -1,0 +1,6 @@
+export interface IBookAdapter<Entity, Dto> 
+{
+  entityToDto(model: Entity): Dto;
+
+  dtoToEntity(dto: Dto): Promise<Entity>;
+}
