--- src/confi/db-config.ts
@@ -1,22 +1,0 @@
-const { Sequelize } = require ('sequelize-typescript');
-import { Books } from "../entit/book-entity" ;
-import { Author } from "../entit/author-entity";
-import { Category } from "../entit/category-entity";
-
-export const sequelize = new Sequelize (
-    {
-        database: 'bms',
-        username: 'postgres',
-        password: 'Odyssey@123',
-        host: 'localhost',
-        port: 5432,
-        dialect: 'postgres',
-        models: [Books, Author, Category],
-        logging: false,
-        define: 
-        {
-            underscored: true,        
-            freezeTableName: true,
-            timestamps: false
-        }
-    });
