--- src/custom-error/custom-erros.ts
@@ -1,11 +1,0 @@
-export class CustomError extends Error {
-  status: number;
-
-  constructor(message: string, status = 400) {
-    super(message);
-    this.name = "CustomError"
-    this.status = status;
-
-    Object.setPrototypeOf(this , new.target.prototype);
-  }
-}
