+++ src/custom-errors/custom-erros.ts
@@ -1,0 +1,11 @@
+export class CustomError extends Error {
+  status: number;
+
+  constructor(message: string, status = 400) {
+    super(message);
+    this.name = "CustomError"
+    this.status = status;
+
+    Object.setPrototypeOf(this , new.target.prototype);
+  }
+}
