--- src/payloa/author-paylaod.ts
@@ -1,5 +1,0 @@
-export class AuhtorDto 
-{
-    authorId!: string;
-    authorName!: string; 
-}
