--- src/payloa/category-payload.ts
@@ -1,5 +1,0 @@
-export class CategoryDto 
-{
-    categoryId!: string;
-    categoryName!: string; 
-}
