+++ src/payload/author-paylaod.ts
@@ -1,0 +1,5 @@
+export class AuhtorDto 
+{
+    authorId!: string;
+    authorName!: string; 
+}
