+++ src/payload/category-payload.ts
@@ -1,0 +1,5 @@
+export class CategoryDto 
+{
+    categoryId!: string;
+    categoryName!: string; 
+}
