--- src/repositor/author-respository.ts
@@ -1,23 +1,0 @@
-import { Service } from "typedi";
-import { Author } from "../entit/author-entity";
-
-@Service()
-export class AuthorRepository {
-
-    async findAuthor(name: string) 
-    {
-        const category = await Author.findOne({ where: { authorName: name } });
-        return category;
-    }
-
-    async createAuthor(data: Author) 
-    {
-        const category = await Author.create(data)
-        return category
-    }
-
-    async findOrCreateAuthor(name: string) 
-    {
-        return await Author.findOrCreate({ where: { authorName: name } })
-    }
-}
