--- src/repositor/category-repository.ts
@@ -1,25 +1,0 @@
-import { Service } from "typedi";
-import { Category } from "../entit/category-entity";
-
-@Service()
-export class CategoryRepository 
-{
-    async findCategory(name: string) 
-    {
-        const category = await Category.findOne({where: {categoryName : name}});
-        return category;
-    }
-
-    async createCategory(data : Category) 
-    {
-        const category = await Category.create(data)
-        return category 
-    }
-
-    async findOrCreateCategory(name : string)
-    {
-        return await Category.findOrCreate({where: {categoryName: name}})
-    }
-
-
-}
