+++ src/repository/author-respository.ts
@@ -1,0 +1,23 @@
+import { Service } from "typedi";
+import { Author } from "../entity/author-entity";
+
+@Service()
+export class AuthorRepository {
+
+    async findAuthor(name: string) 
+    {
+        const category = await Author.findOne({ where: { authorName: name } });
+        return category;
+    }
+
+    async createAuthor(data: Author) 
+    {
+        const category = await Author.create(data)
+        return category
+    }
+
+    async findOrCreateAuthor(name: string) 
+    {
+        return await Author.findOrCreate({ where: { authorName: name } })
+    }
+}
