+++ src/service/author-service.ts
@@ -1,0 +1,21 @@
+import { Service } from "typedi";
+import { AuthorRepository } from "../repository/author-respository";
+import { Author } from "../entity/author-entity";
+
+@Service()
+export class AuthorService
+{
+    private ar: AuthorRepository;
+
+    constructor (authorRepo: AuthorRepository)
+    {
+        this.ar = authorRepo
+    }
+
+    async getOrCreateAuthor (authorName : string): Promise<Author>
+        {
+          const [author] = await this.ar.findOrCreateAuthor(authorName)
+          return author;
+        }
+    
+}
